disable EDN (because graph is too large)
### Out of memory (after formatting EDN)

`FATAL ERROR: Ineffective mark-compacts near heap limit Allocation failed - JavaScript heap out of memory`

If this happens after the step:
*R2G - Formatting EDN (this can take a couple minutes for large graphs),*
then the Roam graph’s EDN is too large for the Action to parse it, because it runs out of memory. Unfortunately, this means you’ll have to disable the EDN backups for it to continue backing up the JSON and Markdown.